# swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: geosys-api
  description: API permettant d'interagir avec GeoSys.
  contact:
    name: Geobase
  version: "1.0"
servers:
- url: http://localhost:8080/geosys-api/v1
paths:
...
  /securite/login:
    get:
      tags:
      - securite
      description: Valide un usager et retourne un jeton (token).
      operationId: get_securite_login
      parameters:
      - name: usager
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: mot_de_passe
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: duree_token
        in: header
        description: Dur�e de vie du token en secondes.
        required: false
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/securite_reponse_login'
              examples:
                exTokenReturn:
                  value:
                    access_token: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE1NzUzMDkxMzAsIm5iZiI6MTU3NTMwOTEzMSwiZXhwIjoxNTc1MzA5MTM1LCJpc3MiOiJOUkNhbiIsInNjb3BlIjoiZ2Vvc3lzOmFkbWluIGdlb3N5czp3cml0ZSBnZW9zeXM6cmVhZCJ9.eFCT0wxG06A7cWmXmLbP5__v-awiYAL_UFrqwVNbdkY
                    token_type: Bearer
                    expired_in: 3600
                    scope:
                    - geosys:read
                    - geosys:write
                    theme:
                    - hydro_50k
                    - corint_250k
                    equipe:
                    - hydro
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnValueBasicErrorMessage'
      x-openapi-router-controller: swagger_server.controllers.securite_controller
...
  /securite/usager/{nom}:
    get:
      tags:
      - securite
      operationId: get_securite_usager_nom
      parameters:
      - name: nom
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/general_message'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/general_message'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/general_message'
      security:
      - geosys_auth:
        - admin
      x-openapi-router-controller: swagger_server.controllers.securite_controller
    delete:
      tags:
      - securite
      operationId: delete_securite_usager_nom
      parameters:
      - name: nom
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/general_message'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/general_message'
      security:
      - geosys_auth:
        - admin
      x-openapi-router-controller: swagger_server.controllers.securite_controller
...
  /securite/usager:
    post:
      tags:
      - securite
      description: Ajouter un nouvel usager de l'API GeoSys.
      operationId: post_securite_usager
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/usager'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/general_message'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/general_message'
      security:
      - geosys_auth:
        - admin
      x-openapi-router-controller: swagger_server.controllers.securite_controller
...
components:
  schemas:
...
    usager:
      title: usager
      type: object
      properties:
        nom:
          type: string
        mot_de_passe:
          type: string
        scopes:
          type: array
          items:
            type: string
        equipes:
          type: array
          items:
            type: string
        themes:
          type: array
          items:
            type: string
        cle_aws_id:
          type: string
...
    securite_reponse_login:
      title: securite_reponse_login
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        scope:
          type: array
          items:
            type: string
        theme:
          type: array
          items:
            type: string
        equipe:
          type: array
          items:
            type: string
      example:
        access_token: access_token
        scope:
        - scope
        - scope
        theme:
        - theme
        - theme
        token_type: token_type
        expires_in: 0
        equipe:
        - equipe
        - equipe
      x-examples: {}
...
    general_message:
      title: general_message
      type: object
      properties:
        message:
          type: string
      example:
        message: message
...
    ReturnValueBasicErrorMessage:
      title: ReturnValue-BasicErrorMessage
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
  securitySchemes:
    geosys_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://geosys/api/oauth/dialog
          scopes:
            write: Ajouter ou modifier l'information
            read: Lire l'information
            delete: Effacer l'information
            admin: Tous les droits
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_geosys_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_geosys_auth